vars and comments
Add variables to colors and commented code that would put the footer at the bottom. The body only begins with the first "real" object. The first margin doesn't count, so it's solved with padding. It could be also solved beginning the page with a height: 0 div.
The problem (reason why it's commented) is that the footer covers the content at the end of the page, because it's position (commented) is absolute at the bottom. I don't wanna put a margin/padding with the exact height of the footer. I want it to be dynamic. adapt at the end of the content correctly, as it would do if its position were static.